Move 'dropTable' into testing module.
